???, also worked on makeGuess (still needs lots of work)
For some reason my JUNIT tests are throwing a FileNotFoundException even though they have not changed. In fact, the driver does find the files.

Also worked on makeGuess, but its tough. Lots to do to make it functional. Currently making a recursive function for the letters that we know are in the word, but do not know its placements. Currently is able to get  a list of those letters ready to be passed to the recursive method, but thats it really.