Driver overhaul
Driver now asks player if they would like to play or if the would like to see the computer play. Does allow for players to play, but the computer option is not written yet. Does allow for multiple games.

Does print the game after every guess in a manner similar to wordle; however, it still does not work properly.

Example: answer is CAPER
when typing PIPER it marks both P's, one as correct and in correct position and one as correct and not in correct position. This should not mark the second P.

Also added a lot of TODOs.