A lot of stuff
Changed structure of score(). now goes through playerGuess(). playerGuess checks for valid input and uses completionState() to check the state of the game. Score now returns a string called score that is the score of that guess.

Changed print statements in letter and game. Much cleaner and easier to read.

Added new functionality to the Driver. Player can now interrupt with an intrusive command that returns data about the game. is meant to be used for development.